requirements.txt
pycryptodomex
PyQt5